feat: Add car, ferry, and passenger domain modules
This commit adds the car, ferry, and passenger domain modules to the project. Each module includes submodules for models, ports, and services. The models define the data structures for cars, ferries, and passengers. The ports define the service interfaces for creating cars, ferries, and passengers. The services implement the service interfaces for creating cars, ferries, and passengers.
