+++ producer/src/lib/domain/ferry/models.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, serde::Serialize)]
+pub struct Ferry {
+    pub id: String,
+    pub name: String,
+    pub capacity: i32,
+}
