+++ producer/src/lib/domain/ferry/services.rs
@@ -1,0 +1,17 @@
+use super::{models::Ferry, ports::FerryService};
+use fake::{faker::company::fr_fr::CompanyName, Fake};
+
+#[derive(Clone, Default)]
+pub struct FerryServiceImpl;
+
+impl FerryService for FerryServiceImpl {
+    async fn create(&self, capacity: i32) -> anyhow::Result<Ferry> {
+        let ferry = Ferry {
+            id: uuid::Uuid::new_v4().to_string(),
+            name: CompanyName().fake(),
+            capacity,
+        };
+
+        Ok(ferry)
+    }
+}
