+++ producer/src/lib/domain/passenger.rs
@@ -1,6 +1,3 @@
-pub mod car;
-pub mod ferry;
 pub mod models;
-pub mod passenger;
 pub mod ports;
 pub mod services;
