+++ producer/src/lib/domain/services.rs
@@ -1,69 +1,13 @@
+use super::car::ports::CarService;
 use std::sync::Arc;
 
-use fake::faker::automotive::fr_fr::LicencePlate;
-use fake::faker::lorem::fr_fr::Word;
-use fake::faker::name::fr_fr::{FirstName, LastName};
-use fake::{faker::company::fr_fr::CompanyName, Fake};
-use rand::Rng;
+use super::ferry::ports::FerryService;
 
-use super::{models::Car, ports::CarService};
+use super::passenger::ports::PassengerService;
 use super::{models::DataSet, ports::DataSetService};
-use super::{models::Ferry, ports::FerryService};
-use super::{models::Passenger, ports::PassengerService};
-
-#[derive(Clone, Default)]
-pub struct FerryServiceImpl;
-
-#[derive(Clone, Default)]
-pub struct CarServiceImpl;
-
-#[derive(Clone, Default)]
-pub struct PassengerServiceImpl;
 
 #[derive(Clone, Default)]
 pub struct DataSetServiceImpl;
-
-impl FerryService for FerryServiceImpl {
-    async fn create(&self, capacity: i32) -> anyhow::Result<Ferry> {
-        let ferry = Ferry {
-            id: uuid::Uuid::new_v4().to_string(),
-            name: CompanyName().fake(),
-            capacity,
-        };
-
-        Ok(ferry)
-    }
-}
-
-impl CarService for CarServiceImpl {
-    async fn create(&self) -> anyhow::Result<Car> {
-        let car = Car {
-            id: uuid::Uuid::new_v4().to_string(),
-            licence_plate: LicencePlate().fake(),
-            brand: Word().fake(),
-            color: Word().fake(),
-            capacity: 5,
-        };
-
-        Ok(car)
-    }
-}
-
-impl PassengerService for PassengerServiceImpl {
-    async fn create(&self, car_id: Option<String>, ferry_id: String) -> anyhow::Result<Passenger> {
-        let mut rng = rand::thread_rng();
-        let passenger = Passenger {
-            id: uuid::Uuid::new_v4().to_string(),
-            car_id,
-            ferry_id,
-            firstname: FirstName().fake(),
-            lastname: LastName().fake(),
-            sex: rng.gen_bool(0.5),
-        };
-
-        Ok(passenger)
-    }
-}
 
 impl DataSetService for DataSetServiceImpl {
     async fn create_data_set<F: FerryService, C: CarService, P: PassengerService>(
@@ -106,9 +50,9 @@ impl DataSetService for DataSetServiceImpl {
         }
 
         Ok(DataSet {
-          ferry,
-          cars,
-          passengers,
-      })
+            ferry,
+            cars,
+            passengers,
+        })
     }
 }
